HPCDATAMGM-1471: Add ability to perform soft deletes

+++ src/hpc-server/hpc-bus-service-impl/src/main/java/gov/nih/nci/hpc/bus/impl/HpcDataManagementBusServiceImpl.java
@@ -86,6 +86,7 @@ import gov.nih.nci.hpc.domain.model.HpcDistinguishedNameSearch;
 import gov.nih.nci.hpc.domain.model.HpcDistinguishedNameSearchResult;
 import gov.nih.nci.hpc.domain.model.HpcRequestInvoker;
 import gov.nih.nci.hpc.domain.model.HpcSystemGeneratedMetadata;
+import gov.nih.nci.hpc.domain.user.HpcAuthenticationType;
 import gov.nih.nci.hpc.domain.user.HpcNciAccount;
 import gov.nih.nci.hpc.domain.user.HpcUserRole;
 import gov.nih.nci.hpc.dto.datamanagement.HpcArchiveDirectoryPermissionsRequestDTO;
@@ -521,7 +522,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 	}
 
 	@Override
-	public void deleteCollection(String path, Boolean recursive) throws HpcException {
+	public void deleteCollection(String path, Boolean recursive, Boolean force) throws HpcException {
 		// Input validation.
 		if (StringUtils.isEmpty(path)) {
 			throw new HpcException("Null / empty path", HpcErrorType.INVALID_REQUEST_INPUT);
@@ -581,7 +582,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 		try {
 			if (recursive) {
 				// Delete all the data objects in this hierarchy first
-				deleteDataObjectsInCollections(path);
+				deleteDataObjectsInCollections(path, force);
 			}
 
 			dataManagementService.delete(path, false);
@@ -1310,7 +1311,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 	}
 
 	@Override
-	public HpcDataObjectDeleteResponseDTO deleteDataObject(String path) throws HpcException {
+	public HpcDataObjectDeleteResponseDTO deleteDataObject(String path, Boolean force) throws HpcException {
 		// Input validation.
 		if (StringUtils.isEmpty(path)) {
 			throw new HpcException("Null / empty path", HpcErrorType.INVALID_REQUEST_INPUT);
@@ -1333,18 +1334,20 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 		}
 
 		// Validate the invoker is the owner of the data object.
-		HpcPermission permission = dataManagementService.getDataObjectPermission(path).getPermission();
-		if (!permission.equals(HpcPermission.OWN)) {
-			throw new HpcException(
-					"Data object can only be deleted by its owner. Your permission: " + permission.value(),
-					HpcRequestRejectReason.DATA_OBJECT_PERMISSION_DENIED);
+		HpcRequestInvoker invoker = securityService.getRequestInvoker();
+		if(!invoker.getAuthenticationType().equals(HpcAuthenticationType.SYSTEM_ACCOUNT)) {
+			HpcPermission permission = dataManagementService.getDataObjectPermission(path).getPermission();
+			if (!permission.equals(HpcPermission.OWN)) {
+				throw new HpcException(
+						"Data object can only be deleted by its owner. Your permission: " + permission.value(),
+						HpcRequestRejectReason.DATA_OBJECT_PERMISSION_DENIED);
+			}
 		}
 
 		// If this is a GroupAdmin, then ensure that:
 		// 1. The file is less than 90 days old
 		// 2. The invoker uploaded the data originally
 
-		HpcRequestInvoker invoker = securityService.getRequestInvoker();
 		if (!registeredLink && HpcUserRole.GROUP_ADMIN.equals(invoker.getUserRole())) {
 			Calendar cutOffDate = Calendar.getInstance();
 			cutOffDate.add(Calendar.DAY_OF_YEAR, -90);
@@ -1383,8 +1386,8 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 			dataObjectDeleteResponse.setLinksDeleteStatus(true);
 		}
 
-		// Delete the file from the archive (if it's archived and not a link).
-		if (!registeredLink) {
+		// Delete the file from the archive (if it's archived and not a link and it is a hard delete).
+		if (!registeredLink && force) {
 			if (!abort) {
 				switch (systemGeneratedMetadata.getDataTransferStatus()) {
 				case ARCHIVED:
@@ -1407,7 +1410,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 		}
 
 		// Remove the file from data management.
-		if (!abort) {
+		if (!abort && force) {
 			try {
 				dataManagementService.delete(path, false);
 				dataObjectDeleteResponse.setDataManagementDeleteStatus(true);
@@ -1418,6 +1421,17 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 				dataObjectDeleteResponse.setMessage(e.getMessage());
 			}
 		} else {
+			if (!abort) {
+				try {
+					dataManagementService.softDelete(path, Optional.of(false));
+					dataObjectDeleteResponse.setDataManagementDeleteStatus(true);
+					dataObjectDeleteResponse.setArchiveDeleteStatus(true);
+				} catch (HpcException e) {
+					logger.error("Failed to soft delete file from datamanagement", e);
+					dataObjectDeleteResponse.setDataManagementDeleteStatus(false);
+					dataObjectDeleteResponse.setMessage(e.getMessage());
+				}
+			}
 			dataObjectDeleteResponse.setDataManagementDeleteStatus(false);
 		}
 
@@ -1760,6 +1774,55 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 		return bulkMoveResponse;
 	}
 
+	@Override
+	public void recoverDataObject(String path) throws HpcException {
+		// Input validation.
+		if (StringUtils.isEmpty(path)) {
+			throw new HpcException("Empty path in recover request: [path: " + path
+					+ "]", HpcErrorType.INVALID_REQUEST_INPUT);
+		}
+
+		HpcDataObject dataObject = dataManagementService.getDataObject(path);
+
+		// Validate the data object exists.
+		if (dataObject == null) {
+			throw new HpcException("Data object doesn't exist: " + path, HpcErrorType.INVALID_REQUEST_INPUT);
+		}
+
+		// Get the metadata for this data object.
+		HpcMetadataEntries metadataEntries = metadataService.getDataObjectMetadataEntries(path);
+		HpcSystemGeneratedMetadata systemGeneratedMetadata = metadataService
+				.toSystemGeneratedMetadata(metadataEntries.getSelfMetadataEntries());
+
+		if (systemGeneratedMetadata.getDataTransferStatus() == null) {
+			throw new HpcException("Unknown data transfer status", HpcErrorType.UNEXPECTED_ERROR);
+		}
+
+		dataManagementService.recover(path, Optional.of(false));
+
+	}
+	
+	@Override
+	public void recoverCollection(String path) throws HpcException {
+		// Input validation.
+		if (StringUtils.isEmpty(path)) {
+			throw new HpcException("Empty path in recover request: [path: " + path
+					+ "]", HpcErrorType.INVALID_REQUEST_INPUT);
+		}
+
+		// Validate the collection exists.
+		HpcCollection collection = dataManagementService.getCollection(path, true);
+		if (collection == null) {
+			throw new HpcException("Collection doesn't exist: " + path, HpcErrorType.INVALID_REQUEST_INPUT);
+		}
+
+		// Recover the collection.
+		recoverDataObjectsFromCollections(path);
+		dataManagementService.delete(path, false);
+
+
+	}
+	
 	// ---------------------------------------------------------------------//
 	// Helper Methods
 	// ---------------------------------------------------------------------//
@@ -1984,7 +2047,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 	 * @param path The path at the root of the hierarchy to delete from.
 	 * @throws HpcException if it failed to delete any object in this collection.
 	 */
-	private void deleteDataObjectsInCollections(String path) throws HpcException {
+	private void deleteDataObjectsInCollections(String path, Boolean force) throws HpcException {
 
 		HpcCollectionDTO collectionDto = getCollectionChildren(path);
 
@@ -1993,7 +2056,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 			if (!CollectionUtils.isEmpty(dataObjects)) {
 				// Delete data objects in this collection
 				for (HpcCollectionListingEntry entry : dataObjects) {
-					deleteDataObject(entry.getPath());
+					deleteDataObject(entry.getPath(), force);
 				}
 			}
 
@@ -2002,7 +2065,7 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 				// Recursively delete data objects from this sub-collection and
 				// it's sub-collections
 				for (HpcCollectionListingEntry entry : subCollections) {
-					deleteDataObjectsInCollections(entry.getPath());
+					deleteDataObjectsInCollections(entry.getPath(), force);
 				}
 			}
 		}
@@ -3179,5 +3242,36 @@ public class HpcDataManagementBusServiceImpl implements HpcDataManagementBusServ
 		resultHpcPermForColl.setCollectionPath(path);
 		resultHpcPermForColl.setPermission(hsPerm.getPermission());
 		return resultHpcPermForColl;
+	}
+	
+	/**
+	 * Recursively recover all the data objects from the specified collection and
+	 * from it's sub-collections.
+	 *
+	 * @param path The path at the root of the hierarchy to recover from.
+	 * @throws HpcException if it failed to recover any object in this collection.
+	 */
+	private void recoverDataObjectsFromCollections(String path) throws HpcException {
+
+		HpcCollectionDTO collectionDto = getCollectionChildren(path);
+
+		if (collectionDto.getCollection() != null) {
+			List<HpcCollectionListingEntry> dataObjects = collectionDto.getCollection().getDataObjects();
+			if (!CollectionUtils.isEmpty(dataObjects)) {
+				// Delete data objects in this collection
+				for (HpcCollectionListingEntry entry : dataObjects) {
+					recoverDataObject(entry.getPath());
+				}
+			}
+
+			List<HpcCollectionListingEntry> subCollections = collectionDto.getCollection().getSubCollections();
+			if (!CollectionUtils.isEmpty(subCollections)) {
+				// Recursively delete data objects from this sub-collection and
+				// it's sub-collections
+				for (HpcCollectionListingEntry entry : subCollections) {
+					recoverDataObjectsFromCollections(entry.getPath());
+				}
+			}
+		}
 	}
 }
